[BE] Refector: API 표준화 (S13P21D103-9)
* refactor: WebSocket 연결 성공 시 메시지 전송 로직을 SocketResponseDispatcher에서 분리

* refector: friend 소켓 타입명 수정

* fix: SocketType 조회 시 null 방지

* refector: friend 에러코드 수정

* refactor: 응답 유형명 수정

* refector: Friend 에러코드 표준화

* fix: UNKNOWN 오류 오탈자 수정

* refactor: 응답 코드 표준화

* refactor: 요청 파싱 중 발생하는 오류 핸들링

* refector: Friend 에러코드 경로 수정 및 표준화

* refector: Auth 에러코드 경로 수정

* refactor: 수신 유형과 송신 유형을 SockeType으로 표준화

* feat: dream에서 발생하는 오류 표준화

* feat: 소켓 응답에 응답 유형 추가

* feat: 소켓 요청 디스패처에 RequestDispatcher 추가

* feat: 소켓 응답 유형 정의

* refactor: 소켓 디스패처에서 동적으로 클래스를 매핑하도록 수정

+++ api-server/src/main/java/com/ggumtle/ggumtle/dream/application/DreamPartyService.java
@@ -1,5 +1,6 @@
 package com.ggumtle.ggumtle.dream.application;
 
+import com.ggumtle.ggumtle.common.SocketType;
 import com.ggumtle.ggumtle.dream.application.command.AcceptPartyInvitationCommand;
 import com.ggumtle.ggumtle.dream.application.command.CreatePartyCommand;
 import com.ggumtle.ggumtle.dream.application.command.InvitePartyCommand;
@@ -10,6 +11,8 @@ import com.ggumtle.ggumtle.dream.domain.PartyParticipant;
 import com.ggumtle.ggumtle.dream.domain.PartyInvitation;
 import com.ggumtle.ggumtle.dream.persistence.PartyInvitationRepository;
 import com.ggumtle.ggumtle.dream.persistence.PartyParticipantRepository;
+import com.ggumtle.ggumtle.exception.GgumtleException;
+import com.ggumtle.ggumtle.exception.code.DreamErrorCode;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
@@ -28,7 +31,7 @@ public class DreamPartyService {
 
     public CreatePartyResult createParty(CreatePartyCommand command) {
         if (partyParticipantRepository.existsById(command.memberId())) {
-            throw new RuntimeException("이미 방에 속해 있습니다");
+            throw new GgumtleException(DreamErrorCode.ALREADY_IN_PARTY);
         }
 
         PartyParticipant partyParticipant = new PartyParticipant(command.memberId(), UUID.randomUUID().toString(), true);
@@ -39,7 +42,9 @@ public class DreamPartyService {
 
     public InvitePartyResult inviteParty(InvitePartyCommand command) {
         PartyParticipant inviter = partyParticipantRepository.findById(command.requesterId())
-                .orElseThrow(() -> new RuntimeException("속한 방이 없습니다"));
+                .orElseThrow(() -> new GgumtleException(
+                        DreamErrorCode.NOT_FOUND_PARTY,
+                        "현재 사용자가 속한 파티가 없어 다른 사용자를 파티에 초대할 수 없습니다"));
 
         PartyInvitation partyInvitation = PartyInvitation.builder()
                 .id(UUID.randomUUID().toString())
@@ -52,17 +57,23 @@ public class DreamPartyService {
         List<Long> memberIds = new ArrayList<>();
         memberIds.add(partyInvitation.getInviteeId());
         List<PartyParticipant> participants = partyParticipantRepository.findAllByPartyId(inviter.getPartyId());
-        participants.stream().forEach(participant -> memberIds.add(participant.getMemberId()));
+        participants.forEach(participant -> memberIds.add(participant.getMemberId()));
 
         return new InvitePartyResult(memberIds, partyInvitation.getId());
     }
 
     public AcceptPartyInvitationResult acceptPartyInvitation(AcceptPartyInvitationCommand command) {
+        if (partyParticipantRepository.existsById(command.requesterId())) {
+            throw new GgumtleException(DreamErrorCode.ALREADY_IN_PARTY);
+        }
+
         PartyInvitation partyInvitation = partyInvitationRepository.findById(command.invitationId())
-                .orElseThrow(() -> new RuntimeException("존재하거나 이미 처리된 초대입니다"));
+                .orElseThrow(() -> new GgumtleException(
+                        DreamErrorCode.NOT_FOUND_INVITATION,
+                        "존재하지 않거나 이미 수락된 초대입니다"));
 
         if (!partyInvitation.getInviteeId().equals(command.requesterId())) {
-            throw new RuntimeException("본인이 받은 파티 초대만 수락할 수 있습니다");
+            throw new GgumtleException(DreamErrorCode.NOT_MY_INVITATION);
         }
 
         partyInvitationRepository.delete(partyInvitation);
